feat(GoalAchievement): add remove criteria action
fix(GoalAchievement): review comments
fix(GoalSuggesstions): add query limit results
