+++ src/components/CriteriaForm/CriteriaForm.i18n/en.json
@@ -1,9 +1,9 @@
 {
     "Weight is required": "",
     "Weight must be integer": "",
-    "Weight": "",
+    "Weight": "No more than {upTo}",
     "Criteria or Goal": "",
-    "Add achivement criteria": "",
+    "Add achievement criteria": "",
     "Weight must be in range": "Weight must be between 1 and {upTo}",
     "Add": ""
 }
