+++ src/components/GoalCriteria/GoalCriteria.i18n/index.ts
