+++ src/components/GoalCriteria/GoalCriteria.i18n/ru.json
