+++ src/hooks/useCriteriaResource.ts
@@ -1,19 +1,12 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback } from 'react';
 
 import { trpc } from '../utils/trpcClient';
 import { AddCriteriaScheme, RemoveCriteriaScheme, UpdateCriteriaScheme } from '../schema/criteria';
 
-export const useCriteriaResource = (invalidateFn: () => Promise<any>) => {
-    const [query, setQuery] = useState('');
+export const useCriteriaResource = (invalidateFn: () => Promise<unknown>) => {
     const add = trpc.goal.addCriteria.useMutation();
     const toggle = trpc.goal.updateCriteriaState.useMutation();
     const remove = trpc.goal.removeCriteria.useMutation();
-
-    const goals = trpc.goal.suggestions.useQuery(query, {
-        enabled: query.length > 2,
-        staleTime: 0,
-        cacheTime: 0,
-    });
 
     const onAddHandler = useCallback(
         async (val: AddCriteriaScheme) => {
@@ -22,6 +15,7 @@ export const useCriteriaResource = (invalidateFn: () => Promise<any>) => {
         },
         [add, invalidateFn],
     );
+
     const onToggleHandler = useCallback(
         async (val: UpdateCriteriaScheme) => {
             await toggle.mutateAsync(val);
@@ -29,6 +23,7 @@ export const useCriteriaResource = (invalidateFn: () => Promise<any>) => {
         },
         [invalidateFn, toggle],
     );
+
     const onRemoveHandler = useCallback(
         async (val: RemoveCriteriaScheme) => {
             await remove.mutateAsync(val);
@@ -37,17 +32,9 @@ export const useCriteriaResource = (invalidateFn: () => Promise<any>) => {
         [invalidateFn, remove],
     );
 
-    useEffect(() => {
-        if (query === '') {
-            goals.remove();
-        }
-    }, [goals, query]);
-
     return {
         onAddHandler,
         onToggleHandler,
         onRemoveHandler,
-        updateSuggestionQuery: setQuery,
-        goals,
     };
 };
