+++ src/schema/schema.i18n/ru.json
@@ -13,6 +13,6 @@
     "Goal's description is required": "",
     "Goal's description must be a string": "",
     "Goal's project or team are required": "",
-    "Criteria Weight is required": "",
-    "Criteria Weight must be longer than 1 symbol": ""
+    "Criteria weight is required": "",
+    "Criteria weight must be longer than 1 symbol": ""
 }
